requirements.txt
Pillow>=9.0
pillow-heif>=0.12.0
